+++ Modules/Project.AllPlatforms.Utility/Project.AllPlatforms.Utility.psm1
@@ -47,6 +47,11 @@ if ($Develop)
 	Write-Debug -Message "[$ThisModule] VerbosePreference is $VerbosePreference"
 	Write-Debug -Message "[$ThisModule] InformationPreference is $InformationPreference"
 }
+else
+{
+	# Everything is default except InformationPreference should be enabled
+	$InformationPreference = "Continue"
+}
 
 # Includes
 . $PSScriptRoot\External\Get-TypeName.ps1