+++ Test/Project.Windows.ProgramInfo/Show-Table.ps1
@@ -1,0 +1,72 @@
+
+<#
+MIT License
+
+Project: "Windows Firewall Ruleset" serves to manage firewall on Windows systems
+Homepage: https://github.com/metablaster/WindowsFirewallRuleset
+
+Copyright (C) 2019, 2020 metablaster zebal@protonmail.ch
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+#>
+
+#
+# Unit test for Show-Table
+#
+. $PSScriptRoot\..\..\Config\ProjectSettings.ps1
+
+# Check requirements for this project
+Import-Module -Name $ProjectRoot\Modules\Project.AllPlatforms.System
+Test-SystemRequirements
+
+# Includes
+. $PSScriptRoot\ContextSetup.ps1
+Import-Module -Name $ProjectRoot\Modules\Project.AllPlatforms.Logging
+Import-Module -Name $ProjectRoot\Modules\Project.AllPlatforms.Test @Logs
+Import-Module -Name $ProjectRoot\Modules\Project.Windows.ProgramInfo @Logs
+Import-Module -Name $ProjectRoot\Modules\Project.AllPlatforms.Utility @Logs
+
+# Ask user if he wants to load these rules
+Update-Context $TestContext $($MyInvocation.MyCommand.Name -replace ".{4}$") @Logs
+if (!(Approve-Execute @Logs)) { exit }
+
+Start-Test
+
+New-Test "-UserProfile switch Fill table with Greenshot"
+Initialize-Table @Logs
+Update-Table "Greenshot" -UserProfile @Logs
+Show-Table
+
+New-Test "Failure Test"
+Initialize-Table @Logs
+Update-Table "Failure" -UserProfile @Logs
+Show-Table
+
+New-Test "Test multiple paths"
+Initialize-Table @Logs
+Update-Table "Visual Studio" -UserProfile @Logs
+Show-Table
+
+New-Test "-Executables switch - Fill table with PowerShell"
+Initialize-Table @Logs
+Update-Table "PowerShell.exe" -Executables
+Show-Table
+
+Update-Logs
+Exit-Test